requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
httpx
pydantic
python-multipart

